use Dir.glob instead of git ls-files in gemspec
Moneta is packaged in Debian, but only at version 1.0. Updating to the
latest version leads to build failures because we do *not* have Git
installed in the build tree: we instead copy a git archive to the
build servers. Details of the build errors are here:

https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1010286#12

I'm not so familiar with the Ruby build process, but from what I
understand, it is actually discouraged to use git in the gemspec in
that way:

https://packaging.rubystyle.guide/#using-git-in-gemspec

I am therefore using the recommended `Dir.glob()` instead of
`git ls-files`.